+++ packages/eslint-config-cnp/src/node.ts
@@ -16,105 +16,16 @@ const config: Linter.Config = {
   parserOptions: {
     ecmaVersion: 2018,
     sourceType: 'module',
-    ecmaFeatures: {
-      jsx: true,
-    },
 
     // typescript-eslint specific options
     warnOnUnsupportedTypeScriptVersion: true,
   },
-  plugins: ['@typescript-eslint', 'cnp', 'import', 'json', 'prettier'],
+  plugins: ['cnp', 'import', 'json', 'prettier'],
 
   extends: [
     'eslint:recommended',
     'plugin:cnp/recommended',
     'plugin:json/recommended-with-comments',
-  ],
-
-  overrides: [
-    {
-      files: ['*.ts?(x)'],
-      extends: [
-        'plugin:@typescript-eslint/eslint-recommended',
-        'plugin:@typescript-eslint/recommended',
-        'plugin:@typescript-eslint/recommended-requiring-type-checking',
-      ],
-      rules: {
-        '@typescript-eslint/array-type': [
-          'error',
-          {
-            default: 'array',
-            readonly: 'array',
-          },
-        ],
-        camelcase: 'off',
-        '@typescript-eslint/naming-convention': [
-          'error',
-          {
-            selector: 'default',
-            format: ['camelCase'],
-          },
-          {
-            selector: ['enum', 'enumMember'],
-            format: ['PascalCase'],
-          },
-          {
-            selector: 'interface',
-            format: ['PascalCase'],
-            custom: {
-              regex: '^I[A-Z]',
-              match: false,
-            },
-          },
-          {
-            selector: 'memberLike',
-            modifiers: ['private'],
-            format: ['camelCase'],
-            leadingUnderscore: 'require',
-          },
-          {
-            selector: 'parameter',
-            format: ['camelCase'],
-            leadingUnderscore: 'forbid',
-          },
-          {
-            selector: 'parameter',
-            modifiers: ['unused'],
-            format: ['camelCase'],
-            leadingUnderscore: 'require',
-          },
-          {
-            selector: 'typeLike',
-            format: ['PascalCase'],
-          },
-          {
-            selector: 'typeParameter',
-            format: ['PascalCase'],
-            prefix: ['T'],
-          },
-          {
-            selector: 'variable',
-            format: ['camelCase', 'UPPER_CASE'],
-          },
-          {
-            selector: 'variable',
-            modifiers: ['global', 'const'],
-            types: ['array', 'boolean', 'number', 'string'],
-            format: ['UPPER_CASE'],
-          },
-        ],
-        '@typescript-eslint/no-unused-vars': [
-          'error',
-          {
-            args: 'after-used',
-            argsIgnorePattern: '^_',
-            destructuredArrayIgnorePattern: '^_',
-            varsIgnorePattern: '^_',
-            ignoreRestSiblings: true,
-          },
-        ],
-      },
-    },
   ],
 
   rules: {